Move files to helpers forlder
