+++ src/helpers/products.ts
