+++ src/view-logic/commonComponentsRendering.ts
@@ -1,7 +1,7 @@
 import Store from "../classes/Store";
 import asideNavBar from "../components/asideNavBar";
-import burgerButtonComponent from "../components/burgerButton";
-import createNewProductButton from "../components/createNewProductButton";
+import burgerButtonComponent from "../helpers/burgerButton";
+import createNewProductButton from "../helpers/createNewProductButton";
 import footerRendering from "../components/footer";
 import headerRendering from "../components/header";
 import PrivilegeEnum from "../enums/PrivilegeEnum";
